Added Git file generation back and custom info endpoint when generation is enable
