+++ service/src/main/scala/za/co/absa/logingw/rest/config/GitConfig.scala
@@ -1,0 +1,76 @@
+/*
+ * Copyright 2023 ABSA Group Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package za.co.absa.logingw.rest.config
+
+import org.springframework.boot.context.properties.{ConfigurationProperties, ConstructorBinding}
+import javax.annotation.PostConstruct
+import java.text.SimpleDateFormat
+import java.io.{PrintWriter, IOException}
+import scala.util.Try
+
+@ConstructorBinding
+@ConfigurationProperties(prefix = "logingw.rest.config")
+case class GitConfig(generateGitProperties: Boolean) {
+
+  @PostConstruct
+  def init(): Unit = {
+    if(generateGitProperties)
+      GitPropertiesGenerator.generateGitProperties()
+  }
+}
+
+case class GitProperties(branch: String, commitId: String, commitTime: String)
+
+object GitPropertiesHolder {
+  var gitProperties: GitProperties = _
+}
+object GitPropertiesGenerator {
+  def generateGitProperties(): Unit = {
+    val gitPropertiesFile = "service\\src\\main\\resources\\git.properties"
+    val dateFormat = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ")
+    val gitBranch = getGitOutput("git rev-parse --abbrev-ref HEAD")
+    val gitCommitId = getGitOutput("git rev-parse HEAD")
+    val gitCommitTime = dateFormat.format(getGitOutput("git show -s --format=%ct HEAD").map(_.toLong * 1000).getOrElse(0L))
+    val content =
+      s"""|git.branch=${gitBranch.getOrElse("unknown")}
+          |git.commit.id=${gitCommitId.getOrElse("unknown")}
+          |git.commit.time=$gitCommitTime
+          |""".stripMargin
+
+    GitPropertiesHolder.gitProperties = GitProperties(gitBranch.getOrElse("unknown"), gitCommitId.getOrElse("unknown"), gitCommitTime)
+    val writer = new PrintWriter(gitPropertiesFile)
+
+    try {
+      writer.println(content)
+    } catch {
+      case ex: IOException =>
+        println(s"Failed to write git.properties file: ${ex.getMessage}")
+    } finally {
+      writer.close()
+    }
+  }
+
+  def getGitOutput(command: String): Option[String] = {
+    Try(sys.process.Process(command).lineStream_!.headOption)
+      .recover {
+        case ex: IOException =>
+          println(s"Failed to execute Git command: $command")
+          None
+      }
+      .getOrElse(None)
+  }
+}
